chore(pseudorand): move from ioutil to testutil
